# perf: use batch loading for BasicUserDetails