--- a/apps/backend/src/config/Tokens.ts
+++ b/apps/backend/src/config/Tokens.ts
@@ -51,4 +51,5 @@ export const Tokens = {
   ProposalInternalCommentsDataSource: Symbol(
     'ProposalInternalCommentsDataSource'
   ),
+  BasicUserDetailsLoader: Symbol('BasicUserDetailsLoader'),
 };