--- a/apps/backend/src/config/dependencyConfigSTFC.ts
+++ b/apps/backend/src/config/dependencyConfigSTFC.ts
@@ -51,6 +51,7 @@ import {
   getStfcDataRow,
   populateStfcRow,
 } from '../factory/xlsx/stfc/StfcFapDataRow';
+import BasicUserDetailsLoader from '../loaders/BasicUserDetailsLoader';
 import { SkipAssetRegistrar } from '../services/assetRegistrar/skip/SkipAssetRegistrar';
 import { configureSTFCEnvironment } from './stfc/configureSTFCEnvironment';
 import { Tokens } from './Tokens';
@@ -118,3 +119,5 @@ mapValue(Tokens.ConfigureEnvironment, configureSTFCEnvironment);
 mapValue(Tokens.ConfigureLogger, () => setLogger(new ConsoleLogger()));
 
 mapClass(Tokens.DownloadService, StfcDownloadService);
+
+mapClass(Tokens.BasicUserDetailsLoader, BasicUserDetailsLoader);