--- a/apps/backend/src/context/index.ts
+++ b/apps/backend/src/context/index.ts
@@ -1,3 +1,4 @@
+import BasicUserDetailsLoader from '../loaders/BasicUserDetailsLoader';
 import UsersLoader from '../loaders/UsersLoader';
 import PDFServices from '../middlewares/factory/factoryServices';
 import { Sdk } from '../middlewares/graphqlClient';
@@ -119,6 +120,7 @@ interface ResolverContextServices {
 }
 interface ResolverContextLoader {
   user: UsersLoader;
+  basicUser: BasicUserDetailsLoader;
 }
 
 export interface BasicResolverContext {