--- a/apps/backend/src/datasources/UserDataSource.ts
+++ b/apps/backend/src/datasources/UserDataSource.ts
@@ -12,6 +12,7 @@ export interface UserDataSource {
   addUserRole(args: AddUserRoleArgs): Promise<boolean>;
   createInviteUser(args: CreateUserByEmailInviteArgs): Promise<number>;
   getBasicUserInfo(id: number): Promise<BasicUserDetails | null>;
+  getBasicUsersInfo(ids: readonly number[]): Promise<BasicUserDetails[]>;
   getBasicUserDetailsByEmail(
     email: string,
     role?: UserRole