<!--- You can leave blank or remove sections which aren't necessary for the PR. -->

## Description

Expands the user batch loading from #469 to apply to BasicUserDetails as well, since the basic variant is quite widely used, and thus causes a lot of calls to the slow external user service at STFC. By combining multiple external calls into one we'll have to pay the time penalty of making network calls fewer times, and when that's ~100ms for each call in our prod environment, it'll add up to quite a bit of user-data-heavy requests.

<!--- Describe your changes in detail -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested

I made 3 local page loads of access panel 11's assignments, recording the time taken by the main 3 calls when starting with an empty users cache. Due to react's dev setup making calls twice, this results in 6 calls for each. For those runs I also tracked the number of UOWS calls made. All times are in milliseconds

Method | Call time (current client) |   |   |   |   |   | Call time (with batching) |   |   |   |   |   | Average time saved | Percentage improvement
---|---|---|---|---|---|---|---|---|---|---|---|---|---|---
getFap | 87 | 89 | 83 | 83 | 104 | 125 | 66 | 68 | 54 | 55 | 73 | 77 | 29.666 | 31%
getFapMembers | 716 | 735 | 704 | 721 | 674 | 675 | 446 | 564 | 589 | 577 | 622 | 642 | 130.833 | 18%
getFapProposals | 711 | 741 | 702 | 725 | 645 | 674 | 538 | 562 | 593 | 581 | 612 | 644 | 111.333 | 16 %

Attempt | UOWS calls - no batching | UOWS calls - batching
---|---|---
1 | 71 | 38
2 | 86 | 48
3 | 72 | 49

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Fixes

<!--- Does this fix a user story, if so add a reference here -->
Closes https://github.com/UserOfficeProject/issue-tracker/issues/1197

## Changes

<!--- What types of changes does your code introduce? In what place? -->

## Depends on

<!--- Does this PR depend on another PR that should be merged first or at the same time -->
https://github.com/UserOfficeProject/mock-uows/pull/19

While the test suite as a whole is passing here, some tests in isolation would have failed without this update to the mock uows, which has already merged. The initial passing state was I believe due to requests for individual users being made first in other tests and then cached in the backend. Without that, any test that first requests multiple users with batch loading would have failed as the mock server would only give one user's details.

## Tests included/Docs Updated?

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->

- [ ] I have added tests to cover my changes.
- [ ] All relevant doc has been updated