@ACLay 2025-02-12
Seems like it. I guess the batching might be done at a more local level within each query, like if we're loading proposals on the panel one by one, it's probably not the fastest to batch up and only request the reviewers at the very end. Plus there's likely going to be a lot of repeats in the reviewers requested in the panel, so a lot of requests are already be dealt with and reduced in the initial unbatched case by the caching.