@ACLay 2025-02-28
I'd been hoping for more impact to this, but since it would only really affect queries when user cache is empty, I suppose little movement in the bulk of the stats should be expected. Still, looks like a nice improvement for the slowest running queries.

Latency charts from dev for the 2 week window either side of this change going in: 
![image](https://github.com/user-attachments/assets/57710330-f65d-4f36-803d-94d65d6a4fd2)